+++ src/Menu/MenuBuy.cpp
@@ -61,7 +61,7 @@ void MenuBuy::menu_print_state(size_t state)
 	{
 		menu_highlight_on();
 	}
-	menu_print(2, state + 4, menu_get_string(state)); // Start at row 4, indent from left border
+	menu_print(2, static_cast<int>(state) + 4, menu_get_string(state)); // Start at row 4, indent from left border
 	if (currentState == state)
 	{
 		menu_highlight_off();
@@ -87,7 +87,7 @@ void MenuBuy::draw()
 	
 	// Title centered on screen
 	size_t title_x = (menu_width - 9) / 2; // Center "Buy Items" (9 chars)
-	mvwprintw(menuWindow, 1, title_x, "Buy Items");
+	mvwprintw(menuWindow, 1, static_cast<int>(title_x), "Buy Items");
 	
 	// Instructions
 	mvwprintw(menuWindow, 2, 2, "Use UP/DOWN or W/S to navigate, ENTER to buy, ESC to exit");
